Remove queue_t
An attempt to remove the Queue Structure previously declared by
making the head an element holding the size of the queue.
In q_new, a memory block size of an int is allocated and assigned to
q->value. It should be treated as an int* and casted into int* before
performing any operations related to the size of the queue.

It may not be the best practice, but it should keep q_size at constant
time without creating another structure just to store the size.